Create pct_sim() and replace word_similar() unusual calc
Percent similarity in word_similar() was calculated as the ratio
of the lenght of input intersection over the length of union of
those inputs. It was designed that way to handle differences in
length between inputs, but was probably flawed. With that approach
the similarity of 1:2 to 2:3 was 33.33% instead of 50% as would
be expected.

Will need to rerun & review the exploration notebook to determine
if this makes a difference in the conclusions.