Create str_compare_all()
To apply nearly any combination of string and word tokenization
transformations and report their similarity. The report is in the
format:

string:<str_transforms>;wordToken[percent similarity]:<word_transforms>

It works on a single small test case. Needs more formal testing.
It may also need to be modified for speed.