Update columns added in auto_review_df()
Dropped curator_match, which was from first approach to curate
labels and is now part of that obsoleted approach. Other than
that specific dataset it is otherwise blank and there is no plan
to use it, so it makes more sense to make the exception there
than have it as a rule.

Added empty final_reviewer column to encourage recording the
person who made the final review.

Updated documentation.