Create script to finalize translation review of disorders
Includes version of Disease Ontology that physical disorder text
was extracted from.

This script:
- Generates a column with comparison between final translation
and prior translations (only for labels)
- Adds predicates for all translated text (using original data
from DO)
- Standardizes data to format envisioned to live in
src/ontology/translations dir of HumanDiseaseOntology repo
(may still need modification)
- Converts that standardized data to a robot template necessary
for generating a release
    * NOTE: The release rules in the HumanDiseaseOntology repo
      does not currently support converting the standardized data
      in src/ontology/translations to robot templates (technical
      issues remain unsolved).

*** This still lacks the code required to compare the source_text
from the ontology with the source_text in these files to
identify where it has changed. This is a critical feature to
ensure translations remain accurate. ***