Update preprocess script with synonym validation
The preprocessing script seemed to be complete and without error
but it seemed worthwhile to add a quick validation to compare
class IDs between English and Spanish translation data to ensure
they did, in fact, match up.

This very limited check identified issues with the data, with rows
9377-11648 having mismatched IDs.

It will be too difficult to programmatically identify what the
issue(s) may be, so the English and Spanish translation data will
be compared manually in a Google sheet to determine if it is
usable.