Fix rate_comparison() calculations
PROBLEMS:
1. Decimals in percent word similarity were being ignored.
2. Word tokenization was deducted with token_wt, as expected,
but was additionally being treated as a word transformation
resulting in an additional word_wt deduction.
3. Character transformations were ignored when word tokenization
was used to create the rating.

FIXES:
1. Allowed decimals in percent word similarity for rating.
2. Updated wordToken so it is not treated as a word transformation
and, additionally, updated the default token_wt to better match
previous behavior (ratings seemed reasonable for labels at least).
3. Updated calculations to include both character & word
transformations when tokenization is used. Just like token_wt and
word_wt, chr_wt is scaled by the percent word similarity. All 3
weights have reduced impact as percent similarity drops.

RESULT:
The net result of these updates is a change in mean rating by
about -0.0038, with a lower bound of about -0.042 and an upper
bound of 0.008 for definitions and a more conservative change for
labels (-0.025, 0.007; mean = -0.0001).

The definition rating weights may need to be optimized.