Append data in script to finalize "remaining" diseases
This ensures only translations newly marked as final are added,
and preserves previously finalized translations without
modification.

Appending seems less ideal (compared to overwriting) but an
unexpected request to update translation production files with
each DO release made it necessary to TEMPORARLIY implement
(it was more likely to be correct with less effort). It might be
better to compare the latest review data with the data in the
existing, production translation TSV for changes and to
control the order records are written in to make it easier to
review diffs.